feat: Refactor login components and added new Font Style
