+++ src/components/Login/LoginButton.jsx
@@ -1,7 +1,7 @@
-import "../styles/LoginPage.css";
+import style from "./styles/LoginPage.module.css";
 const LoginButton = () => {
   return (
-    <button type="button" className="btn">
+    <button type="button" className={style.btn}>
       Login
     </button>
   );
