+++ src/components/Login/PasswordInput.jsx
@@ -1,0 +1,8 @@
+import style from "./styles/LoginPage.module.css";
+const PasswordInput = () => {
+  return (
+    <input type="password" className={style.input} placeholder="Password" />
+  );
+};
+
+export default PasswordInput;
