+++ src/components/Login/UsernameInput.jsx
@@ -1,0 +1,6 @@
+import style from "./styles/LoginPage.module.css";
+const UsernameInput = () => {
+  return <input type="text" className={style.input} placeholder="Username" />;
+};
+
+export default UsernameInput;
