+++ src/components/Login/styles/LoginPage.module.css
@@ -2,10 +2,9 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 100vh; /* Full viewport height */
-  background-color: var(
-    --color-bg-dark-primary
-  ); /* Background color for the container */
+  height: 100vh;
+  margin: 0;
+  background-color: var(--color-bg-dark-primary);
 }
 
 .form {
@@ -13,49 +12,29 @@
   justify-content: center;
   align-items: center;
   transform-style: preserve-3d;
-  transition: all 1s ease;
+  transition: all 3s ease;
 }
 
-.form .form_front {
+.form_container {
   display: flex;
-  flex-direction: column;
+  flex-direction: column; /* Stack elements vertically */
   justify-content: center;
   align-items: center;
-  gap: 20px;
-  position: absolute;
-  backface-visibility: hidden;
-  padding: 65px 45px;
-  border-radius: 15px;
-  box-shadow: inset 2px 2px 10px var(--color-bg-dark-secondary),
-    inset -1px -1px 5px rgba(255, 255, 255, 0.6);
-}
-
-.form .form_back {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-  position: absolute;
-  backface-visibility: hidden;
-  transform: rotateY(-180deg);
-  padding: 65px 45px;
-  border-radius: 15px;
-  box-shadow: inset 2px 2px 10px rgba(0, 0, 0, 1),
-    inset -1px -1px 5px rgba(255, 255, 255, 0.6);
+  gap: 20px; /* Spacing between elements */
 }
 
 .form_details {
-  font-size: 25px;
+  font-size: 35px;
   font-weight: 600;
-  padding-bottom: 10px;
-  color: var(--color-text-dark-primary);
+  padding-bottom: 12px;
+  color: var(--color-text-dark-secondary);
 }
 
 .input {
   width: 245px;
   min-height: 45px;
-  color: var(--color-bg-dark-secondary);
+  font-size: 18px;
+  color: var(--color-text-dark-primary);
   outline: none;
   transition: 0.35s;
   padding: 0px 7px;
@@ -67,7 +46,7 @@
 }
 
 .input::placeholder {
-  color: var(--color-text-dark-secondary);
+  color: var(--color-text-dark-primary);
 }
 
 .input:focus.input::placeholder {
@@ -91,7 +70,7 @@
   box-shadow: 6px 6px 10px rgba(0, 0, 0, 1),
     1px 1px 10px rgba(255, 255, 255, 0.6);
   color: var(--color-text-dark-primary);
-  font-size: 15px;
+  font-size: 18px;
   font-weight: bold;
   transition: 0.35s;
 }
@@ -110,7 +89,7 @@
 }
 
 .form .switch {
-  font-size: 13px;
+  font-size: 23px;
   color: var(--color-text-dark-primary);
 }
