--- src/components/PasswordInput.jsx
@@ -1,6 +1,0 @@
-import "../styles/LoginPage.css";
-const PasswordInput = () => {
-  return <input type="password" className="input" placeholder="Password" />;
-};
-
-export default PasswordInput;
