--- src/components/UsernameInput.jsx
@@ -1,6 +1,0 @@
-import "../styles/LoginPage.css";
-const UsernameInput = () => {
-  return <input type="text" className="input" placeholder="Username" />;
-};
-
-export default UsernameInput;
