+++ src/index.css
@@ -1,7 +1,5 @@
-*{
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
+@import url("https://fonts.googleapis.com/css2?family=Noto+Sans:ital,wght@0,100..900;1,100..900&family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Urbanist:ital,wght@0,100..900;1,100..900&display=swap");
 
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
+* {
+  font-family: "Open Sans";
 }
