+++ src/pages/LoginPage.jsx
@@ -1,14 +1,13 @@
-import UsernameInput from "../components/UsernameInput";
-import PasswordInput from "../components/PasswordInput";
-import LoginButton from "../components/LoginButton";
-import "../styles/LoginPage.css";
-
+import UsernameInput from "../components/Login/UsernameInput";
+import PasswordInput from "../components/Login/PasswordInput";
+import LoginButton from "../components/Login/LoginButton";
+import style from "../components/Login/styles/LoginPage.module.css";
 const LoginPage = () => {
   return (
-    <div className="container">
-      <form className="form">
-        <div className="form_front">
-          <div className="form_details">Login</div>
+    <div className={style.container}>
+      <form className={style.form}>
+        <div className={style.form_container}>
+          <div className={style.form_details}>Login</div>
           <UsernameInput />
           <PasswordInput />
           <LoginButton />
